refactor: prep work for tensorzero-providers crate extraction
Move shared types and utilities out of tensorzero-core into
tensorzero-inference-types, in preparation for extracting
crates/tensorzero-core/src/providers/ into a standalone
tensorzero-providers crate in a follow-up PR.

Moved (re-exported from core for backward compatibility):

- `InferenceProvider`, `WrappedProvider`, `TensorZeroEventError`
  from `tensorzero-core/src/inference/mod.rs` to a new
  `tensorzero-inference-types::provider_trait` module. The trait
  is the linchpin every provider implements; relocating it lets
  the providers crate live without depending on tensorzero-core.

- `BatchRequestRow` and `UnparsedBatchRequestRow` (plus their
  `sqlx::FromRow` impl for `PgRow`) into
  `tensorzero-inference-types`. The orphan rule means the FromRow
  impl had to follow the type.

- `deprecation_warning`, `is_mock_mode`, `get_mock_provider_api_base`
  into a new `tensorzero-inference-types::utils` module.

Also exposes `tensorzero-inference-types::serde_helpers` and adds
`deserialize_json_string` to it (needed by the moved BatchRequestRow).

Adds `async-trait` and `reqwest-sse-stream` as deps of
`tensorzero-inference-types`.

No behavior change. `cargo check --all-targets --all-features`
and `cargo clippy --all-targets --all-features -- -D warnings`
both pass.
