+++ crates/tensorzero-inference-types/Cargo.toml
@@ -6,8 +6,10 @@ rust-version.workspace = true
 license.workspace = true
 
 [dependencies]
+async-trait.workspace = true
 derive_builder.workspace = true
 futures.workspace = true
+reqwest-sse-stream.workspace = true
 mime.workspace = true
 mime_guess = "2.0.5"
 rust_decimal = { workspace = true }
